+++ app/javascript/components/Reservations/index.jsx
@@ -18,7 +18,7 @@ export default class Reservations extends React.Component {
       rooms: this.props.rooms,
       startDate: moment(new Date).format('DD.MM.YYYY'),
       endDate: moment(new Date).add(1, 'days').format('DD.MM.YYYY'),
-      floor: '1',
+      floor: 1,
       name: '',
       phone: '',
       places: 1,
@@ -26,9 +26,6 @@ export default class Reservations extends React.Component {
       editModal: false,
       selectedRoomId: ''
     }
-  }
-
-  componentDidMount() {
   }
 
   handleDateChange = (event, picker) => {
@@ -90,7 +87,7 @@ export default class Reservations extends React.Component {
       }
     }).then((resp) => {
       if (resp.success) {
-        this.setState({rooms: resp.rooms, createModal: false, name: '', phone: '', places: ''})
+        this.setState({rooms: resp.rooms, createModal: false, name: '', phone: '', places: 1})
       } else {
         NotificationManager.error(resp.error, 'Неможливо створити');
       }
@@ -152,17 +149,19 @@ export default class Reservations extends React.Component {
   }
 
   deleteReservation = (id) => {
-    $.ajax({
-      url: `/hotels/${this.state.hotelId}/reservations/${id}.json`,
-      type: 'DELETE',
-      data: {
-        floor: this.state.floor,
-        start_date: this.state.startDate,
-        end_date: this.state.endDate
-      }
-    }).then((resp) => {
-      this.setState({rooms: resp.rooms})
-    });
+    if (confirm('Видалити дане бронювання?')) {
+      $.ajax({
+        url: `/hotels/${this.state.hotelId}/reservations/${id}.json`,
+        type: 'DELETE',
+        data: {
+          floor: this.state.floor,
+          start_date: this.state.startDate,
+          end_date: this.state.endDate
+        }
+      }).then((resp) => {
+        this.setState({rooms: resp.rooms, editModal: false})
+      });
+    }
   }
 
   render() {
@@ -175,15 +174,18 @@ export default class Reservations extends React.Component {
             <DateRangePicker onApply={this.handleDateChange} startDate={this.state.startDate} endDate={this.state.endDate}>
               <div className='row'>
                 <div className='col-lg-6'>
+                  <label>Початок заїзду</label>
                   <input type="text" className='form-control' value={this.state.startDate}/>
                 </div>
                 <div className='col-lg-6'>
+                  <label>Кінець заїзду</label>
                   <input type="text" className='form-control' value={this.state.endDate}/>
                 </div>
               </div>
             </DateRangePicker>
           </div>
-          <div className='col-lg-4'>
+          <div className='col-lg-2'>
+            <label>Поверх</label>
             <select className='form-control' value={this.state.floor} onChange={(e) => this.handleFloorChange(e.target.value)}>
               { [...Array(parseInt(this.state.floors, 10))].map((e,i) =>
                 <option key={i} value={i+1}>Поверх {i+1}</option>
@@ -191,17 +193,26 @@ export default class Reservations extends React.Component {
             </select>
           </div>
         </div>
+        <hr/>
         { Object.keys(this.state.rooms).length > 0 ?
           <div className='row'>
             { Object.keys(this.state.rooms).map((id) => {
               return (
-                <div key={id} className={`col-lg-3 room ` + this.roomColor(id)}>
-                  <p>Номер {this.state.rooms[id].number}</p>
-                  <p>Місць {this.state.rooms[id].places}</p>
+                <div key={id} className='col-lg-12 room'>
+                  <h4 className={'float-right ' + this.roomColor(id)}>
+                    <strong>
+                      { this.roomColor(id) === 'red' && 'Заброньований'}
+                      { this.roomColor(id) === 'yellow' && 'Частково заброньований'}
+                      { this.roomColor(id) === 'green' && 'Вільний'}
+                    </strong>
+                  </h4>
+                  <h3><b>Номер {this.state.rooms[id].number}</b></h3>
+                  <p>Кількість місць: {this.state.rooms[id].places}</p>
+                  { this.roomColor(id) === 'yellow' && <p>Кількість заброньованих місць: {this.state.rooms[id].booked}</p>}
                   { this.roomColor(id) != 'red' &&
-                    <button className='btn-dark' onClick={() => this.handleModal('createModal', id)}>Create</button>}
+                    <button className='btn-dark' onClick={() => this.handleModal('createModal', id)}>Створити резервування</button>}
                   { this.roomColor(id) != 'green' &&
-                    <button className='btn-info' onClick={() => this.handleModal('editModal', id)}>Edit</button>}
+                    <button className='btn-info' onClick={() => this.handleModal('editModal', id)}>Редагувати бронь</button>}
                 </div>
               )})}
           </div>
@@ -210,31 +221,37 @@ export default class Reservations extends React.Component {
         { this.state.createModal &&
           <Modal isOpen={this.state.createModal} toggle={() => this.handleModal('createModal', '')}>
             <div className='reservation-form'>
-              <label>Name</label>
+              <h6 className='text-center'>Створення нового бронювання {`від ${this.state.startDate} до ${this.state.endDate}`}</h6>
+              <label>Ім'я</label>
               <input type='text' className='form-control' value={this.state.name} onChange={(e) => this.handleInputChange('name', e.target.value)} />
-              <label>Phone</label>
+              <label>Телефон</label>
               <input type='text' className='form-control' value={this.state.phone} onChange={(e) => this.handleInputChange('phone', e.target.value)} />
-              <label>Places</label>
+              <label>Кількість місць</label>
               <select className='form-control' value={this.state.places} onChange={(e) => this.handleInputChange('places', e.target.value)}>
                 { [...Array(parseInt(this.state.rooms[this.state.selectedRoomId].places, 10))].map((e,i) =>
                   <option key={i} value={i+1}>{i+1}</option>
                 )}
               </select>
-              <button className='btn btn-block' onClick={this.handleSubmitReservation}>Submit</button>
+              <button className='btn btn-block reservation-btn' onClick={this.handleSubmitReservation}>Створити</button>
             </div>
           </Modal>}
         { this.state.editModal &&
           <Modal isOpen={this.state.editModal} toggle={() => this.handleModal('editModal', '')}>
+            <div className='reservation-form-top'>
+              <h5 className='text-center'>Номер {this.state.rooms[this.state.selectedRoomId].number}</h5>
+              <h6 className='text-center'>Поточні бронювання {`від ${this.state.startDate} до ${this.state.endDate}`}</h6>
+              <hr/>
+            </div>
             { Object.keys(this.state.rooms[this.state.selectedRoomId].reservations).map((r, i) => {
               const reservation = this.state.rooms[this.state.selectedRoomId].reservations[r]
               return (
                 <div className='reservation-form' key={i}>
-                  <i className='fa fa-trash-o' onClick={() => this.deleteReservation(r)} />
-                  <label>Name</label>
+                  <i className='fa fa-trash-o float-right' onClick={() => this.deleteReservation(r)} />
+                  <label>Ім'я</label>
                   <input type='text' className='form-control' value={reservation.name} onChange={(e) => this.handleReservationChange(r, 'name', e.target.value)} />
-                  <label>Phone</label>
+                  <label>Телефон</label>
                   <input type='text' className='form-control' value={reservation.phone} onChange={(e) => this.handleReservationChange(r, 'phone', e.target.value)} />
-                  <label>Places</label>
+                  <label>Кількість місць</label>
                   <select className='form-control' value={reservation.places} onChange={(e) => this.handleReservationChange(r, 'places', e.target.value)}>
                     { [...Array(parseInt(this.state.rooms[this.state.selectedRoomId].places, 10))].map((e,i) =>
                       <option key={i} value={i+1}>{i+1}</option>
@@ -242,7 +259,10 @@ export default class Reservations extends React.Component {
                   </select>
                 </div>
               )})}
-            <button className='btn btn-block' onClick={this.handleSubmitEditReservation}>Submit</button>
+              <div className='reservation-btn-div'>
+                <hr/>
+                <button className='btn btn-block reservation-btn' onClick={this.handleSubmitEditReservation}>Редагувати</button>
+              </div>
           </Modal>}
       </div>
     );
