+++ app/javascript/components/Rooms/Room.jsx
@@ -62,23 +62,25 @@ export default class Room extends React.Component {
   }
 
   deleteReservation = () => {
-    $.ajax({
-      url: `/hotels/${this.state.hotelId}/reservations/${this.state.selectedReservation.id}.json`,
-      type: 'DELETE',
-      data: {
-        from_calendar: true,
-        date: this.state.selectedMonth
-      }
-    }).then((resp) => {
-      this.setState({
-        ...this.state,
-        editModal: false,
-        room: {
-          ...this.state.room,
-          reservations: resp.reservations
+    if (confirm('Видалити дане бронювання?')) {
+      $.ajax({
+        url: `/hotels/${this.state.hotelId}/reservations/${this.state.selectedReservation.id}.json`,
+        type: 'DELETE',
+        data: {
+          from_calendar: true,
+          date: this.state.selectedMonth
         }
-      });
-    })
+      }).then((resp) => {
+        this.setState({
+          ...this.state,
+          editModal: false,
+          room: {
+            ...this.state.room,
+            reservations: resp.reservations
+          }
+        });
+      })
+    }
   }
 
   handleSubmitEditReservation = () => {
@@ -197,7 +199,7 @@ export default class Room extends React.Component {
           newReservation: {
             name: '',
             phone: '',
-            places: '',
+            places: 1,
             startDate: moment(new Date).format('DD.MM.YYYY'),
             endDate: moment(new Date).add(1, 'days').format('DD.MM.YYYY'),
             roomId: Object.keys(this.state.rooms)[0]
@@ -216,12 +218,15 @@ export default class Room extends React.Component {
     return (
       <div className="container">
         <NotificationContainer/>
-        <select className='form-control' value={this.state.room.id} onChange={(e) => this.handleRoomChange(e.target.value)}>
-          { Object.keys(this.state.rooms).map((id, i) =>
-            <option key={i} value={id}>Номер {this.state.rooms[id].number}</option>
-          )}
-        </select>
-        <button className='btn-default' onClick={() => this.handleModal('createModal')}>New Reservation</button>
+        <div className='calendar-top'>
+          <select className='form-control' value={this.state.room.id} onChange={(e) => this.handleRoomChange(e.target.value)}>
+            { Object.keys(this.state.rooms).map((id, i) =>
+              <option key={i} value={id}>Номер {this.state.rooms[id].number} - Кількість місць: {this.state.rooms[id].places}</option>
+            )}
+          </select>
+          <button className='btn-dark' onClick={() => this.handleModal('createModal')}>Створити нове бронювання</button>
+        </div>
+        <hr/>
         <div className='room-calendar'>
           <createSlot waitForOutlet />
           <BigCalendar
@@ -239,17 +244,17 @@ export default class Room extends React.Component {
         { this.state.createModal &&
           <Modal isOpen={this.state.createModal} toggle={() => this.handleModal('createModal')}>
             <div className='reservation-form'>
-              <label>Name</label>
+              <label>Ім'я</label>
               <input type='text' className='form-control' value={this.state.newReservation.name} onChange={(e) => this.handleNewReservationChange('name', e.target.value)} />
-              <label>Phone</label>
+              <label>Телефон</label>
               <input type='text' className='form-control' value={this.state.newReservation.phone} onChange={(e) => this.handleNewReservationChange('phone', e.target.value)} />
-              <label>Places</label>
+              <label>Кількість місць</label>
               <select className='form-control' value={this.state.newReservation.places} onChange={(e) => this.handleNewReservationChange('places', e.target.value)}>
                 { [...Array(parseInt(this.state.room.places, 10))].map((e,i) =>
                   <option key={i} value={i+1}>{i+1}</option>
                 )}
               </select>
-              <label>Dates</label>
+              <label>Дати</label>
               <DateRangePicker
                 onApply={this.handleNewReservationDateChange}
                 startDate={this.state.newReservation.startDate}
@@ -263,24 +268,25 @@ export default class Room extends React.Component {
                   </div>
                 </div>
               </DateRangePicker>
-              <button className='btn btn-block' onClick={this.handleSubmitReservation}>Submit</button>
+              <hr/>
+              <button className='btn btn-block reservation-btn' onClick={this.handleSubmitReservation}>Створити</button>
             </div>
           </Modal>}
         { this.state.editModal &&
           <Modal isOpen={this.state.editModal} toggle={() => this.handleModal('editModal')}>
             <div className='reservation-form'>
-              <i className='fa fa-trash-o' onClick={this.deleteReservation} />
-              <label>Name</label>
+              <i className='fa fa-trash-o float-right' onClick={this.deleteReservation} />
+              <label>Ім'я</label>
               <input type='text' className='form-control' value={this.state.selectedReservation.title} onChange={(e) => this.handleReservationChange('title', e.target.value)} />
-              <label>Phone</label>
+              <label>Телефон</label>
               <input type='text' className='form-control' value={this.state.selectedReservation.phone} onChange={(e) => this.handleReservationChange('phone', e.target.value)} />
-              <label>Places</label>
+              <label>Кількість місць</label>
               <select className='form-control' value={this.state.selectedReservation.places} onChange={(e) => this.handleReservationChange('places', e.target.value)}>
                 { [...Array(parseInt(this.state.room.places, 10))].map((e,i) =>
                   <option key={i} value={i+1}>{i+1}</option>
                 )}
               </select>
-              <label>Dates</label>
+              <label>Дати</label>
               <DateRangePicker onApply={this.handleDateChange} startDate={moment(this.state.selectedReservation.start).format('DD.MM.YYYY')} endDate={moment(this.state.selectedReservation.end).format('DD.MM.YYYY')}>
                 <div className='row'>
                   <div className='col-lg-6'>
@@ -292,7 +298,10 @@ export default class Room extends React.Component {
                 </div>
               </DateRangePicker>
             </div>
-            <button className='btn btn-block' onClick={this.handleSubmitEditReservation}>Submit</button>
+            <div>
+              <hr/>
+              <button className='btn btn-block reservation-btn' onClick={this.handleSubmitEditReservation}>Редагувати</button>
+            </div>
           </Modal>}
       </div>
     );
